% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acronym-package.R
\docType{package}
\name{acronym-package}
\alias{acronym}
\alias{acronym-package}
\title{acronym: Find acronyms in your Word documents}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

It helps finding acronyms in your Word document for making a list of abbreviations.
}
\author{
\strong{Maintainer}: Felipe Mattioni Maturana \email{felipe.mattioni@med.uni-tuebingen.de} (\href{https://orcid.org/0000-0002-4221-6104}{ORCID})

}
\keyword{internal}
