% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acRonym.R
\name{get_acronyms}
\alias{get_acronyms}
\title{Get acronyms from given text}
\usage{
get_acronyms(str)
}
\arguments{
\item{str}{A string containing the acronyms.}
}
\value{
A tibble with columns "Acronym" and "Definition"
}
\description{
Creates a tibble with a list of acronyms defined in `str` along with a guess at their
definition. Acronyms are defined as occurring within parentheses
}
\examples{
\dontrun{
get_acronyms(readLines(file))
}
}
