% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acRonym.R
\name{get_acronyms_dir}
\alias{get_acronyms_dir}
\title{Get acronyms from all files in a directory}
\usage{
get_acronyms_dir(path, extension = "tex")
}
\arguments{
\item{path}{The root folder containing the relevant files}

\item{extension}{The file extension, defaults to "tex"}
}
\value{
A tibble containing columns "Acronym" and "Definition"
}
\description{
Runs \code{get_acronyms} recursively over all specified files in a directory.
}
