% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regex_acronyms}
\alias{regex_acronyms}
\title{Regex for finding acronyms}
\usage{
regex_acronyms(text)
}
\arguments{
\item{text}{The text column retrieved from \code{officer::read_docx() \%>\% officer::docx_summary()}.}
}
\value{
a character list
}
\description{
Regex for finding acronyms
}
